% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fit_functions.R
\name{BCSreg}
\alias{BCSreg}
\title{Box-Cox Symmetric Regression for Positive Data}
\usage{
BCSreg(
  formula,
  data,
  subset,
  na.action,
  family = "NO",
  zeta,
  link = "log",
  sigma.link = "log",
  control = BCSreg.control(...),
  model = TRUE,
  y = TRUE,
  x = FALSE,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model, of type \code{y ~ x} to specify
explanatory variables in the scale (\code{mu}) regression structure only or \code{y ~ x | z}
to also specify explanatory variables in the relative dispersion (\code{sigma}) regression
structure. See details below.}

\item{data, subset, na.action}{arguments controlling formula processing via
\code{\link[stats]{model.frame}}.}

\item{family}{a character that specifies the symmetric generating family of the BCS distribution.
Available options are: \code{"NO"} (default), \code{"ST"}, \code{"LOI"}, \code{"LOII"},
\code{"PE"}, \code{"SN"}, \code{"HP"}, and \code{"SL"}, corresponding to the normal,
Student-\emph{t}, type I logistic, type II logistic, power exponential, sinh-normal,
hyperbolic, and slash distributions, respectively.}

\item{zeta}{strictly positive extra parameter. It must be specified with only one value
in cases where the BCS distribution has an extra parameter. See “Details” below.}

\item{link, sigma.link}{character specification of the link function for the
scale (\code{mu}) and relative dispersion (\code{sigma}) regression structures, respectively.
Currently, \code{"log"} (default), \code{"sqrt"}, \code{"1/mu^2"},
\code{"inverse"}, and \code{"identity"} are supported.}

\item{control}{a list of control parameters passed as arguments for
the \code{\link[stats]{optim}} function specified via \code{\link{BCSreg.control}}.}

\item{model, y, x}{logicals. If \code{TRUE}, the corresponding components of the fit
(model frame, response, and model matrices) are returned.}

\item{...}{arguments passed to \code{\link{BCSreg.control}}.}
}
\value{
The \code{BCSreg} function returns an object of class \code{"BCSreg"},
     which consists of a list with the following components:
 \describe{
    \item{coefficients}{ a list containing the elements \code{"mu"} and
        \code{"sigma"} that consist of the estimates of the coefficients
         associated with the scale and relative dispersion regression structures, respectively.}
    \item{fitted.values}{ a vector with the fitted median responses. Not to be confused with the fitted values for \code{mu}.}
    \item{mu}{ a vector with the fitted scale parameters.}
    \item{sigma}{ a vector with the fitted relative dispersion parameters.}
    \item{lambda}{ the maximum likelihood estimate of the skewness parameter (\code{lambda}), or its fixed value
        specified in the \code{BCSreg} function.}
    \item{zeta}{ the specified value for the extra parameter of the corresponding BCS distribution, if applicable.}
    \item{family}{ the generating family of the fitted BCS distribution.}
    \item{link}{ a list with elements \code{"mu"} and \code{"sigma"} with the
        specified link functions for the \code{mu} and \code{sigma} regression structures, respectively.}
    \item{logLik}{ log-likelihood of the fitted model.}
    \item{vcov}{ asymptotic covariance matrix of the estimators. By default, the asymptotic
        covariance matrix is based on a analytical expression of the observed information matrix.
        It can be obtained numerically based on the Hessian matrix via \code{\link[stats]{optim}})
        if the argument \code{hessian = TRUE} is used in the \code{BCSreg} function.}
    \item{residuals}{ a vector of quantile residuals.}
    \item{pseudo.r.squared}{pseudo R-squared value.}
    \item{Upsilon.zeta}{an overall goodness-of-fit measure.}
    \item{v}{a vector with the \eqn{v(z)} values for all the observations.
        The \eqn{v} function is a weighting function that is involved in the
        parameter estimation process. It depends on the generating density function
        of the corresponding BCS distribution.}
    \item{nobs}{ number of observations.}
    \item{df.null}{ residual degrees of freedom in the null model (constant scale and relative dispersion).}
    \item{df.residual}{ residual degrees of freedom in the fitted model.}
    \item{control}{the control arguments passed to the optim call.}
    \item{start}{a vector with the starting values used in the iterative process.}
    \item{optim}{a list with the output from \code{\link[stats]{optim}}.}
    \item{converged}{logical indicating successful convergence of the iterative process.}
    \item{call}{the original function call.}
    \item{formula}{the formula used.}
    \item{terms}{a list with elements "\code{mu}", "\code{sigma}", and "\code{full}" containing
        the term objects for the respective models.}
    \item{levels}{a list with elements "\code{mu}", "\code{sigma}", and "\code{full}" containing
        the levels of the categorical regressors.}
    \item{contrasts}{a list with elements "\code{mu}" and "\code{sigma}"
        containing the contrasts corresponding to levels from the respective models.}
    \item{model}{the full model frame (if \code{y = TRUE}).}
    \item{y}{the response variable (if \code{y = TRUE}).}
    \item{x}{a list with elements "\code{mu}" and "\code{sigma}" with the model matrices from
        the \code{mu} and \code{sigma} submodels (if \code{x = TRUE}).}
   }
}
\description{
Maximum likelihood estimation in the class of the Box-Cox symmetric (BCS) regression models
    for the analysis of positive data.
}
\details{
The \code{BCSreg} function implements maximum likelihood estimation in the
    class of the BCS regression models for the analysis of positive
    data (Medeiros and Queiroz, 2025). The BCS distributions (Ferrari and Fumes, 2017)
    are a broad class of flexible distributions that achieve different levels of
    skewness and tail-heaviness. See details in \link{bcs}.

    The distributions currently implemented in the \code{BCSreg} package, along
    with their abbreviations used in the \code{family} argument, are listed below:
    \tabular{llc}{
     \bold{Distribution}  \tab \bold{Family abbreviation} \tab \bold{Number of parameters}\cr
     Box-Cox Hyperbolic  \tab \code{"HP"}      \tab  4  \cr
     Box-Cox Type I Logistic  \tab \code{"LOI"}      \tab  3  \cr
     Box-Cox Type II Logistic  \tab \code{"LOII"}      \tab  3  \cr
     Box-Cox Normal  \tab \code{"NO"}      \tab  3  \cr
     Box-Cox Power Exponential  \tab \code{"PE"}      \tab  4  \cr
     Box-Cox Sinh-Normal  \tab \code{"SN"}      \tab  4  \cr
     Box-Cox Slash  \tab \code{"SL"}      \tab  4  \cr
     Box-Cox \emph{t}  \tab \code{"ST"}      \tab  4  \cr
     }

    The BCS distributions have at least three parameters: scale (\code{mu}),
    relative dispersion (\code{sigma}), and skewness (\code{lambda}) parameters.
    Some distributions may also depend on an additional parameter (\code{zeta}),
    such as the Box-Cox \emph{t} and Box-Cox power exponential distributions. The
    BCS distributions reduce to the log-symmetric distributions when \code{lambda}
    is fixed at zero. The Log-symmetric distributions are an important class of
    probability models for positive data, which includes well-known distributions such
    as the log-normal and log-\emph{t} distributions. The \code{BCSreg} function
    allows fitting a log-symmetric regression using \code{lambda = 0} as an
    argument (see \code{\link{BCSreg.control}}).


    The \code{BCSreg} function allows defining a regression structure in the
    \code{mu} and \code{sigma} parameters via the \code{formula} argument. The
    basic formula is of type \code{y ~ x1 + x2 + ... + xp} which specifies the
    regression structure for \code{mu} only. Following the syntax of the
    \code{betareg} package (Cribari-Neto and Zeileis, 2010), the regression structure
    for \code{sigma}, say in terms of \code{s1, s2, ..., sq}, is specified as
    \code{y ~ x1 + x2 + ... + xp | s1 + s2 + ... + sq} using functionalities inherited
    from the \code{\link[Formula]{Formula}} package (Zeileis and Croissant, 2010).
}
\references{
Cribari-Neto F, Zeileis A (2010). Beta Regression in R. \emph{Journal of Statistical
    Software}, \bold{34}, 1-24

    Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
    applications to nutritional data. \emph{AStA Advances in Statistical Analysis}, \bold{101}, 321-344.

    Medeiros, R. M. R., and Queiroz, F. F. (2025). Modeling positive continuous data:
    Box-Cox symmetric regression models and their extensions

    Zeileis A, Croissant Y (2010). Extended Model Formulas in R: Multiple Parts and Multiple
    Responses. \emph{Journal of Statistical Software}, \bold{34}, 1-13.
}
