% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_methods.R
\name{BCSreg_methods}
\alias{BCSreg_methods}
\alias{model.frame.BCSreg}
\alias{model.matrix.BCSreg}
\alias{coef.BCSreg}
\alias{vcov.BCSreg}
\alias{logLik.BCSreg}
\alias{AIC.ugrpl}
\title{Extract Information From a Box-Cox Symmetric Regression Fit}
\usage{
\method{model.frame}{BCSreg}(formula, ...)

\method{model.matrix}{BCSreg}(object, model = c("mu", "sigma", "alpha"), ...)

\method{coef}{BCSreg}(object, model = c("mu", "sigma", "alpha", "full"), ...)

\method{vcov}{BCSreg}(object, model = c("mu", "sigma", "alpha", "full"), ...)

\method{logLik}{BCSreg}(object, ...)

\method{AIC}{ugrpl}(object, ..., k = 2)
}
\arguments{
\item{formula}{a model formula or terms object or an R object.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object of class \code{"BCSreg"}, a result of a call to \link{BCSreg}.}

\item{model}{a character indicating which regression structure should be used.
It can be \code{"mu"} for the scale regression structure, \code{"sigma"} for
the relative dispersion regression structure, or \code{"full"} (when applicable)
for both regression structures.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical Akaike information criteria (AIC).}
}
\description{
Methods for \code{"BCSreg"} objects.
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
