% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_extra_parameter.R
\name{extra.parameter}
\alias{extra.parameter}
\alias{print.extra.parameter}
\alias{plot.extra.parameter}
\title{Select the Extra Parameter of a Box-Cox Symmetric Regression}
\usage{
extra.parameter(
  object,
  family,
  grid = seq(1, 30, 2),
  trace = TRUE,
  plot = TRUE,
  control = BCSreg.control(...),
  ...
)

\method{print}{extra.parameter}(x, ...)

\method{plot}{extra.parameter}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"BCSreg"}, a result of a call to \code{\link{BCSreg}}.}

\item{family}{a character that specifies the symmetric generating family of the
BCS distribution. The options available are: \code{"ST"}, \code{"PE"},
\code{"SN"}, \code{"HP"}, and \code{"SL"}. The \code{"NO"}, \code{"LOI"},
and \code{"LOII"} generating families do not depend on additional parameters.}

\item{grid}{grid of values that will be used to evaluate the Upsilon statistics and the profiled
log-likelihood function.}

\item{trace}{logical; if \code{TRUE}, a summary with the profiled log-likelihood value and the
Upsilon statistics is displayed.}

\item{plot}{logical; if \code{TRUE}, a graph of the Upsilon statistics evaluated in the
considered grid of values is shown.}

\item{control}{a list of control arguments specified via \code{\link{BCSreg.control}}.}

\item{...}{further arguments passed to \code{\link{BCSreg.control}}.}

\item{x}{an object of class \code{"extra.parameter"}.}
}
\value{
An object of class \code{"extra.parameter"}. More specifically, it returns a list in which
    each element consists of the fit of the BCS regression with each value of the extra
    parameter zeta specified in \code{grid}. In addition, it has the elements \code{"logLik"}
    with the vector of log-likelihood values for each fit, \code{"Upsilon"} with the values
    of the Upsilon statistics for each fit, and \code{"grid"} with the specified grid of
    values.

    The \code{print} function summarizes the fits by displaying, for each value in \code{grid},
    the log-likelihood value and the Upsilon statistic. The
    \code{plot} function returns a graph of the Upsilon statistics, highlighting its
    minimum.
}
\description{
Estimation of the extra parameter in a Box-Cox regression fit based on the Upsilon
    goodness-of-fit statistic and profiled likelihood.
}
\examples{
# Examples

}
\references{
Medeiros, R. M. R., and Queiroz, F. F. (2025). Modeling positive continuous data:
    Box-Cox symmetric regression models and their extensions
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
