% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_diagnostic.R
\name{influence}
\alias{influence}
\title{Influence Diagnostics for BCSreg Objects}
\usage{
influence(object, plot = TRUE, ask = grDevices::dev.interactive(), ...)
}
\arguments{
\item{object}{an object of class \code{"BCSreg"}.}

\item{plot}{logical. If \code{plot = TRUE} (default), the plots are shown.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot, if \code{plot = TRUE}.}

\item{...}{currently not used.}
}
\value{
\code{influence} returns a list with three objects:
    \item{case.weights}{The values of \eqn{d_{max}} eigenvector based on case
    weights perturbation scheme (see Medeiros and Queiroz (2025)).}
    \item{totalLI}{The total local influence (see Lesaffre and Verbeke (1998)).}
}
\description{
The \code{influence} function provides two influence measures for Box-Cox symmetric regression
    models.
}
\examples{
## Examples

}
\references{
Lesaffre, E., and Verbeke, G. (1998). Local influence in linear mixed models.
 \emph{Biometrics}, 570--582.

 Medeiros, R. M. R., and Queiroz, F. F. (2025). Modeling positive continuous data:
    Box-Cox symmetric regression models and their extensions
}
\seealso{
\code{\link{BCSreg}}, \code{\link{plot.BCSreg}}
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
