% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{plot.BCSreg}
\alias{plot.BCSreg}
\title{Diagnostic Plots for a Box-Cox Symmetric Regression Fit}
\usage{
\method{plot}{BCSreg}(
  x,
  which = 1:4,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  pch = "+",
  las = 1,
  cex = 0.8,
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"sdlrm"}.}

\item{which}{numeric; if a subset of the plots is required, specify a subset
of the numbers \code{1:6}.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{pch, las, cex, lwd, ...}{graphical parameters (see \code{\link[graphics]{par}})}
}
\value{
\code{plot} method for \code{"\link{BCSreg}"} objects returns six types
    of diagnostic plots.
}
\description{
This function provides plots for diagnostic analysis of a Box-Cox symmetric
    regression fit.
}
\details{
The \code{plot} method for \code{\link{BCSreg}} objects provides six types
    of diagnostic plots in the following order:
    \describe{
        \item{Residuals vs fitted values}{a plot of the residuals
            versus the fitted medians.}
        \item{Residuals vs observation indices.}{an index plot of the residuals
            versus the observation indices.}
        \item{Density plot}{a graph that compares the empirical density of the residuals
            with the density of the standard normal distribution.}
        \item{Normal probability plot}{a normal probability plot of the residuals with a
            confidence region constructed according to Fox (2016) using the
            \code{\link[car]{qqPlot}} function.}
        \item{Fitted vs observed values}{a dispersion diagram of the fitted values
            versus the observed values.}
        \item{Residuals vs v(z) function}{a dispersion diagram of the \eqn{v(z)} function
            versus the residuals. For someBCS models, the \eqn{v(z)} function
            may be interpreted as weights in the estimation process. If \code{family = "NO"},
            the \eqn{v(z)} function is constant.}
     }

     The \code{which} argument can be used to select a subset of the implemented plots.
     Default is \code{which = 1:4}.
}
\examples{
## Examples
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
