% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{residuals}
\alias{residuals}
\alias{residuals.BCSreg}
\title{Extract Residuals for a Box-Cox Symmetric Regression Fit}
\usage{
\method{residuals}{BCSreg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"BCSreg"}, a result of a call to \link{BCSreg}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Function \code{residuals} returns a vector with the quantile residuals
    resulting from a Box-Cox symmetric regression fit.
}
\description{
Extract Residuals for a Box-Cox Symmetric Regression Fit
}
\examples{
## Examples
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
