% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{summary.BCSreg}
\alias{summary.BCSreg}
\alias{print.summary.BCSreg}
\title{Summarizing a Box-Cox Symmetric Regression Fit}
\usage{
\method{summary}{BCSreg}(object, ...)

\method{print}{summary.BCSreg}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{an object of class \code{"BCSreg"}, a result of a call to \link{BCSreg}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.BCSreg"}, a result of a call to \link{summary.BCSreg}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values.}
}
\value{
The function \code{summary.BCSreg} returns an object of class \code{"summary.BCSreg"},
    which consists of a list with the following components:
 \describe{
    \item{call}{the original function call, given in \code{object}.}
    \item{mu}{summary statistics for the scale submodel.}
    \item{sigma}{summary statistics for the relative dispersion submodel.}
    \item{lambda}{summary statistics for the skewness parameter. If this parameter is not
        statistically different from zero, the fitted Box-Cox symmetric (BCS) distribution
        can be reduced to a more parsimonious log-symmetric distribution.}
    \item{alpha}{summary statistics for the zero-adjustment submodel when a zero-adjusted
        model is considered; and \code{NULL}, otherwise.}
    \item{zeta}{the specified value for the extra parameter of the fitted
        BCS distribution, if applicable.}
    \item{family}{the generating family of the fitted BCS distribution.}
    \item{link}{a list with elements \code{"mu"} and \code{"sigma"} with the
        specified link functions for the \code{mu} and \code{sigma} regression
        structures, respectively. If the model is zero-adjusted, the element
        \code{"alpha"} will also be returned with the link function for
        the regression structure of the zero-adjustment parameter.}
    \item{converged}{logical indicating successful convergence of the iterative
        process.}
    \item{iterations}{number of iterations reached until the optimization algorithm converges.}
    \item{logLik}{log-likelihood of the fitted model.}
    \item{df}{number of model parameters}
    \item{residuals}{a vector of quantile residuals.}
    \item{pseudo.r.squared}{pseudo R-squared value.}
    \item{Upsilon.zeta}{an overall goodness-of-fit measure.}
    \item{v}{a vector with the \eqn{v(z)} values for all the observations.
        The \eqn{v} function is a weighting function that is involved in the
        parameter estimation process. It depends on the generating density function
        of the corresponding BCS distribution.}
    \item{AIC, BIC}{Akaike and Bayesian information criteria.}
 }
}
\description{
\code{summary} method for class \code{"BCSreg"}.
}
\examples{
## Examples
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
