% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_methods.R
\name{BCSreg-methods}
\alias{BCSreg-methods}
\alias{model.frame.BCSreg}
\alias{model.matrix.BCSreg}
\alias{coef.BCSreg}
\alias{vcov.BCSreg}
\alias{logLik.BCSreg}
\alias{AIC.ugrpl}
\alias{print.BCSreg}
\alias{summary.BCSreg}
\alias{print.summary.BCSreg}
\title{Extract Information From a Box-Cox Symmetric Regression Fit}
\usage{
\method{model.frame}{BCSreg}(formula, ...)

\method{model.matrix}{BCSreg}(object, model = c("mu", "sigma"), ...)

\method{coef}{BCSreg}(object, model = c("mu", "sigma", "full"), ...)

\method{vcov}{BCSreg}(object, model = c("mu", "sigma", "full"), ...)

\method{logLik}{BCSreg}(object, ...)

\method{AIC}{ugrpl}(object, ..., k = 2)

\method{print}{BCSreg}(x, digits = getOption("digits"), ...)

\method{summary}{BCSreg}(object, ...)

\method{print}{summary.BCSreg}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{formula}{a model formula or terms object or an R object.}

\item{...}{further arguments passed to or from other methods.}

\item{model}{a character indicating which regression structure should be used.
It can be \code{"mu"} for the scale regression structure, \code{"sigma"} for
the relative dispersion regression structure, or \code{"full"} (when applicable)
for both regression structures.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical Akaike information criteria (AIC).}

\item{x, object}{an object of class \code{"BCSreg"}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values. The default, \code{getOption("digits")}.}
}
\description{
Methods for \code{"BCSreg"} and \code{"summary.BCSreg"} objects.
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
