% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{BCSreg-methods}
\alias{BCSreg-methods}
\alias{model.frame.BCSreg}
\alias{model.matrix.BCSreg}
\alias{coef.BCSreg}
\alias{vcov.BCSreg}
\alias{logLik.BCSreg}
\alias{AIC.ugrpl}
\title{Extract Information From a Box-Cox Symmetric Regression Fit}
\usage{
\method{model.frame}{BCSreg}(formula, ...)

\method{model.matrix}{BCSreg}(object, model = c("mu", "sigma", "alpha"), ...)

\method{coef}{BCSreg}(object, model = c("mu", "sigma", "alpha", "full"), ...)

\method{vcov}{BCSreg}(object, model = c("mu", "sigma", "alpha", "full"), ...)

\method{logLik}{BCSreg}(object, ...)

\method{AIC}{ugrpl}(object, ..., k = 2)
}
\arguments{
\item{formula}{a model formula or terms object or an R object.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object of class \code{"BCSreg"}, a result of a call to \link{BCSreg}.}

\item{model}{a character indicating which regression structure should be used.
It can be \code{"mu"} for the scale regression structure, \code{"sigma"} for
the relative dispersion regression structure, or \code{"full"} (when applicable)
for both regression structures.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical Akaike information criteria (AIC).}
}
\value{
\itemize{
\item \code{model.frame} returns a \code{data.frame} containing the variables required
    by \code{formula} and any additional arguments provided via \code{...}.
\item \code{model.matrix} returns the design matrix used in the regression structure,
    as specified by the \code{model} argument.
\item \code{coef} returns a numeric vector of estimated regression coefficients, based
    on the \code{model} argument. If \code{parm = "full"}, it returns a list with the
    components \code{"mu"} and \code{"sigma"}, each containing the corresponding
    coefficient estimates. If the model is zero-adjusted, it will also have a
    \code{"alpha"} component with the estimates of the associated regression coefficients.
\item \code{vcov} returns the asymptotic covariance matrix of the regression coefficients,
    based on the \code{model} argument.
\item \code{logLik} returns the log-likelihood value of the fitted model.
\item \code{AIC} returns a numeric value representing the Akaike Information Criterion
    (AIC), Bayesian Information Criterion, or another criterion, depending on \code{k}.
}
}
\description{
Methods for \code{"BCSreg"} objects.
}
\examples{
## Data set: raycatch (for description, run ?raycatch)
hist(raycatch$cpue, xlab = "Catch per unit effort")
plot(cpue ~ tide_phase, raycatch, pch = 16,
   xlab = "Tide phase", ylab = "Catch per unit effort")
plot(cpue ~ location, raycatch, pch = 16,
   xlab = "Location", ylab = "Catch per unit effort")
plot(cpue ~ max_temp, raycatch, pch = 16,
   xlab = "Maximum temperature", ylab = "Catch per unit effort")

## Fit the Box-Cox normal regression as a reference model
fit <- BCSreg(cpue ~ location + tide_phase |
                location + tide_phase + max_temp, raycatch)

## coef
coef(fit)
coef(fit, model = "sigma")
coef(fit, model = "full")

## vcov
vcov(fit)
vcov(fit, model = "sigma")
vcov(fit, model = "full")

## Log-likelihood value
logLik(fit)

## AIC and BIC
AIC(fit)
AIC(fit, k = log(fit$nobs))

## Model matrices
model.matrix(fit)
model.matrix(fit, model = "sigma")
}
\references{
Ferrari, S. L. P., and Fumes, G. (2017). Box-Cox symmetric distributions and
    applications to nutritional data. \emph{AStA Advances in Statistical Analysis},
    \bold{101}, 321---344.

    Medeiros, R. M. R., and Queiroz, F. F. (2025). Flexible modeling of non-negative continuous
    data: Box-Cox symmetric regression and its zero-adjusted extension.
}
\author{
Francisco F. de Queiroz <\email{felipeq@ime.usp.br}>

Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
