% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCSreg.R
\name{BCSregcontrol}
\alias{BCSregcontrol}
\alias{BCSreg.control}
\title{Auxiliary for Controlling BCS Fitting}
\usage{
BCSreg.control(
  lambda = NULL,
  method = "BFGS",
  maxit = 2000,
  trace = FALSE,
  start = NULL,
  ...
)
}
\arguments{
\item{lambda}{numeric indicating the value of lambda (if \code{NULL}, lambda
will be estimated).}

\item{method}{character specifying the \code{method} argument passed to \code{\link{optim}}.}

\item{maxit, trace, ...}{arguments passed to \code{\link{optim}}}

\item{start}{an optional vector with starting values for median and dispersion
submodels (starting value for lambda must not be included).}
}
\value{
A list with components named as the arguments.
}
\description{
Parameters that control fitting of Box-Cox symmetric regression models using \code{\link{BCSreg}}.
}
\details{
The \code{BCSreg.control} controls the fitting process of Box-Cox models.
    Almost all the arguments are passed directly to \code{\link{optim}}, which
    is used to estimate the parameters. Starting values for median and dispersion
    submodels can be provided via \code{start}. The starting value for lambda is
    zero; that is, the fitting process starts from the log-symmetric regression
    model. If the estimation process is to be performed with a fixed lambda,
    a value must be specified in \code{lambda}. If \code{lambda = 0}, a log-symmetric
    regression model will be estimated.
}
\examples{
2+2=4
}
\seealso{
\code{\link{BCSreg}}
}
