% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_fit_functions.R
\name{BCSregcontrol}
\alias{BCSregcontrol}
\alias{BCSreg.control}
\title{Auxiliary for Controlling BCS Fitting}
\usage{
BCSreg.control(
  lambda = NULL,
  method = "BFGS",
  maxit = 2000,
  hessian = FALSE,
  trace = FALSE,
  start = NULL,
  ...
)
}
\arguments{
\item{lambda}{numeric indicating the value of lambda (if \code{NULL}, lambda
will be estimated).}

\item{method}{character specifying the \code{method} argument passed to \code{\link{optim}}.}

\item{maxit, trace, ...}{arguments passed to \code{\link{optim}}.}

\item{hessian}{logical. Should the numerical Hessian matrix from the \code{optim} output be
used for estimation of the covariance matrix? By default the analytical solution is employed.}

\item{start}{an optional vector with starting values for the regression coefficients
associated with the \code{mu} and \code{sigma} submodels (starting value for the
\code{lambda} parameter must not be included).}
}
\value{
A list with components named as the arguments.
}
\description{
Optimization parameters that control the fitting of Box-Cox symmetric regression
    models using the \code{\link{BCSreg}} function.
}
\details{
The \code{BCSreg.control} controls the fitting process of Box-Cox symmetric models.
    Almost all the arguments are passed directly to \code{\link{optim}}, which
    is used to estimate the parameters. Starting values for the regression coefficients
    associated with the \code{mu} and \code{sigma} submodels can be provided via \code{start}
    argument. The starting value for the skewness parameter \code{lambda} is zero; that is,
    the fitting process starts from the corresponding log-symmetric regression model. If the
    estimation process is to be performed with a fixed \code{lambda}, a value must be specified
    for the \code{lambda} argument. A natural value is \code{lambda = 0}, where a log-symmetric
    regression model will be estimated.
}
\examples{
# Set an example


}
\seealso{
\code{\link{BCSreg}}
}
